weekly-projects/week9-project/requirements.txt
Django==2.2
Faker==1.0.4
IMDbPY==6.6
lxml==4.3.3
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
SQLAlchemy==1.3.2
sqlparse==0.3.0
text-unidecode==1.2
